sqlhist: Do not process previous keys into variables
The algorithm got a bit overzealous and added more variables when it needed
to display variables from the first event, but the kernel parsing does not
work well with variables assigned to other variables. Just use the variables
that are assigned in the first event in the second, and that appears to
work.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>